+++ src/lib.rs
@@ -193,6 +193,14 @@ mod record;
 mod result_code;
 pub mod task;
 mod user;
+pub mod entity;
 
 #[cfg(test)]
 extern crate hex;
+
+
+#[cfg(feature = "derive")]
+extern crate aerospike_derive;
+
+#[cfg(feature = "derive")]
+pub use aerospike_derive::*;