+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+* **New Feature**
+  * Add ```Entity``` trait for more flexibility for change between Models and Records
+  * Add ```EntityClient``` a client wrapper for flexibility
+  * Add ```derive``` feature for auto implement entity traits
 * **Compatibility**
   * Add ```From``` trait for ```Option<T>``` in both direction. by [Ali Mirghasemi](https://github.com/ali-mirghasemi)
   * Add ```From``` trait for string array ```Vec<String>``` in both direction. by [Ali Mirghasemi](https://github.com/ali-mirghasemi)